matttest/requirements.txt
pyyaml==5.3.1
pytest==6.0.2
kubernetes==11.0